Set up stub handler to process multiple events
