+++ cmd/federation-v2-operator/main.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"runtime"
 
-	"github.com/marun/federation-v2-operator/pkg/federation"
+	"github.com/marun/federation-v2-operator/pkg/stub"
 	sdk "github.com/operator-framework/operator-sdk/pkg/sdk"
 	k8sutil "github.com/operator-framework/operator-sdk/pkg/util/k8sutil"
 	sdkVersion "github.com/operator-framework/operator-sdk/version"
@@ -32,6 +32,6 @@ func main() {
 	resyncPeriod := 5
 	logrus.Infof("Watching %s, %s, %s, %d", resource, kind, namespace, resyncPeriod)
 	sdk.Watch(resource, kind, namespace, resyncPeriod)
-	sdk.Handle(federation.NewHandler())
+	sdk.Handle(stub.NewHandler())
 	sdk.Run(context.TODO())
 }
